+++ Assets/Mirror/Runtime/NetworkClient.cs
@@ -450,6 +450,7 @@ namespace Mirror
             if (LogFilter.Debug) Debug.Log("Shutting down client.");
             ClientScene.Shutdown();
             connectState = ConnectState.None;
+            handlers.Clear();
         }
 
         /// <summary>